Implemented lazy loading for the task lists by category
On clicking a category on the left side, the list of tasks is updated to those for that category. This is done via an API request. Once fetched, the list is saved in the task store w/ key as its category. No API request will be repeated for the lists